refactor: split resource store into modules (actions, utils, types, storage)
