+++ data/projects.json
@@ -7,14 +7,14 @@
       "url": "",
       "projectType": "",
       "customer": "",
-      "order": 4
+      "order": 3
     },
     {
-      "name": "Проект 55555",
+      "name": "Проект 555",
       "archived": false,
       "id": 5,
       "url": "teccode.ru",
-      "order": 0,
+      "order": 1,
       "customer": "ООО \"техинвест\"",
       "projectType": "Разработка"
     },
@@ -28,22 +28,23 @@
       "name": "Проект 10",
       "archived": false,
       "id": 7,
-      "order": 1,
-      "url": ""
+      "order": 4,
+      "url": "",
+      "customer": ""
     },
     {
       "name": "Проект 11",
       "archived": false,
       "url": "",
       "id": 8,
-      "order": 2
+      "order": 0
     },
     {
       "name": "Проект 15",
       "archived": false,
       "url": "",
       "id": 9,
-      "order": 3
+      "order": 2
     },
     {
       "name": "Проект 13",
@@ -60,11 +61,22 @@
       "order": 6
     },
     {
-      "name": "Проект 12",
+      "name": "вфывфв",
       "archived": false,
+      "order": 8,
       "url": "",
-      "id": 12,
-      "order": 8
+      "customer": "ООО \"Технокод\"",
+      "projectType": "Разработка",
+      "id": 12
+    },
+    {
+      "name": "1234",
+      "archived": false,
+      "order": 9,
+      "url": "",
+      "customer": "12",
+      "projectType": "34",
+      "id": 13
     }
   ]
 }
