+++ src/stores/resource/constants.ts
@@ -1,0 +1,1 @@
+export const HIDDEN_GROUPS_STORAGE_KEY = 'resource_hidden_group_ids';
